+++ gluuengine/setup/base.py
@@ -273,6 +273,28 @@ class BaseSetup(object):
             hostname = "ldap.{}".format(dns_search.rstrip("."))
         return hostname
 
+    def get_web_cert(self):
+        hostname = self.cluster.ox_cluster_hostname.split(":")[0]
+
+        if not os.path.exists(self.app.config["SSL_CERT_DIR"]):
+            os.makedirs(self.app.config["SSL_CERT_DIR"])
+
+        ssl_cert = os.path.join(self.app.config["SSL_CERT_DIR"], "nginx.crt")
+        ssl_key = os.path.join(self.app.config["SSL_CERT_DIR"], "nginx.key")
+
+        if os.path.exists(ssl_cert) and os.path.exists(ssl_key):
+            # copy cert and key
+            self.logger.debug("copying existing SSL cert")
+            self.docker.copy_to_container(self.container.cid, ssl_cert, "/etc/certs/nginx.crt")
+            self.logger.debug("copying existing SSL key")
+            self.docker.copy_to_container(self.container.cid, ssl_key, "/etc/certs/nginx.key")
+        else:
+            self.gen_cert("nginx", self.cluster.decrypted_admin_pw,
+                          "www-data", "www-data", hostname)
+            # save certs locally, so we can reuse and distribute them
+            self.docker.copy_from_container(self.container.cid, "/etc/certs/nginx.crt", ssl_cert)
+            self.docker.copy_from_container(self.container.cid, "/etc/certs/nginx.key", ssl_key)
+
 
 class OxSetup(BaseSetup):
     def write_salt_file(self):