+++ gluuengine/task/licensewatcher.py
@@ -7,6 +7,7 @@ import logging
 import time
 
 from crochet import run_in_reactor
+from requests.exceptions import ConnectionError
 from twisted.internet.task import LoopingCall
 
 from ..database import db
@@ -61,8 +62,12 @@ class LicenseWatcherTask(object):
             self.logger.info("license key is currently unavailable")
             return
 
-        # get current datetime from license server
-        current_date = retrieve_current_date()
+        try:
+            # get current datetime from license server
+            current_date = retrieve_current_date()
+        except ConnectionError:
+            self.logger.warn("unable to get current date from license server")
+            return
 
         # if license has been already updated within 24 hours,
         # no need to re-populate the license
@@ -113,12 +118,7 @@ class LicenseWatcherTask(object):
                     # if we have specific containers being disabled in node,
                     # try to re-enable the containers
                     self.enable_containers(node, "oxauth")
-
-            # distribute data as we may have new state about containers
-            if worker_nodes:
-                distribute_cluster_data(
-                    self.app.config["SHARED_DATABASE_URI"], self.app,
-                )
+                distribute_cluster_data(self.app, node)
 
     def get_license_key(self):
         with self.app.app_context():